Explicitly sets whether pages are published
The published value in the front matter controls whether to show or hide links in nav-links.html. Running jekyll locally with the --unpublished option will include unpublished items.

Also adds an include which can wrap another page targetted include so long as the targetted page is published. This can be used for links elsewhere.