Uses Jekyll to run Sass
Removed the app.css file as this will now be generated for us.
Moved the app.scss file into the css directory so that Jekyll knows where to place the rendered version.
Config now tells Jekyll where to find the Sass files.
Places a copy of the foundation and motion-ui scss files in to the scss directory.
Tweaks the scss/base/_settings.scss file to fix some errors with units that were reported by Jekyll when it ran Sass.
Adds some Mapbox attribution styling to the buildings-tours scss.